django.contrib.staticfiles support: tests are passed on both django 1.2 and django 1.3.
MARKITUP_MEDIA_URL is now equal to STATIC_URL by default and falls back to
MEDIA_URL if STATIC_URL is not defined (e.g. django 1.2 without django-staticfiles).

If STATIC_URL is defined then it is assumed that either django 1.3 or
django-staticfiles is in use and JQuery is served from STATIC_URL, not
from MEDIA_URL. Django 1.2 prepends MEDIA_URL to relative form media so
JQuery URL is forced to be absolute there and is always prepended with STATIC_URL
if STATIC_URL is defined.